LPS-62912 move classes
